[fix] : 로그 컴포넌트 분리, 스타일 조정
